+++ app/views/admin/index.html.slim
@@ -11,8 +11,8 @@ dl.dl-horizontal
     = link_to 'CSV出力', admin_csvs_path
   dd 各種データをCSV形式でダウンロードします。
   dt
-    = link_to '稼働集計', summary_reports_path
+    = link_to '稼働集計', summary_path
   dd 全員の稼働を集計して表示します。
   dt
-    = link_to '日報未提出一覧', unsubmitted_reports_path
+    = link_to '日報未提出一覧', unsubmitted_path
   dd 日報が未提出である日付をユーザーごとに一覧表示します。
