+++ app/views/reports/summaries/show.html.slim
@@ -1,0 +1,38 @@
+- breadcrumb :summary
+
+h1 稼働集計
+
+= render_flash_message
+
+= form_tag reports_path, method: :get, class: 'form-inline', id: 'summary-render' do
+  .form-group
+    = date_select_ja :reports, :start, default: @date_start
+    | 〜
+    = date_select_ja :reports, :end, default: @date_end
+    | の稼働集計を
+  = submit_tag '表示', class: 'btn btn-default'
+
+= form_tag summary_path(format: :csv), method: :get, id: 'summary-download' do
+  = submit_tag 'CSVダウンロード', class: 'btn btn-primary'
+
+- if @sum.present?
+  .well= "集計期間：#{@date_start.strftime('%Y年%-m月%-d日')} 〜 #{@date_end.strftime('%Y年%-m月%-d日')}"
+
+  table.table.table-bordered.table-striped
+    tr
+      th PJコード
+      th プロジェクト名
+      - @users.each do |user|
+        th= user.name
+    - @sum.each do |sum|
+      tr
+        th= @projects[sum[0]].code
+        th= @projects[sum[0]].name
+        - @users.each do |user|
+          td= sum[1][user.id] if sum[1].has_key?(user.id)
+
+javascript:
+  new ReportSummary({
+    renderForm: 'summary-render',
+    downloadForm: 'summary-download'
+  });
